// src/course/enum.ts
export enum CourseLevel {
  Beginner = "beginner",
  Intermediate = "intermediate",
  Advanced = "advanced",
}

export enum Currency {
  NGN = "NGN",
  USD = "USD",
  EUR = "EUR",
  GBP = "GBP",
}

export enum CourseStatus {
  Draft = "draft",
  Published = "published",
  Archived = "archived",
}

export enum Language {
  English = "english",
  French = "french",
  Spanish = "spanish",
  Yoruba = "yoruba",
  Igbo = "igbo",
  Hausa = "hausa",
}

export enum CurriculumItemType {
  Lecture = "lecture",
  Article = "article",
  Resource = "resource",
  Quiz = "quiz",
  Assignment = "assignment",
}

export enum EnrollmentStatus {
  Active = "active",
  Completed = "completed",
  Dropped = "dropped",
}

export enum Category {
  WebDevelopment = "web_development",
  MobileDevelopment = "mobile_development",
  Design = "design",
  UIUX = "ui_ux",
  DataScience = "data_science",
  MachineLearning = "machine_learning",
  CloudComputing = "cloud_computing",
  AWS = "aws",
  FullStack = "full_stack",
}